# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/get-all-admins": {
      "get": {
        "description": "Every admin and super admin, newest first, including deactivated ones (check `isActive`). No request body and no parameters. Password hashes are never included.",
        "operationId": "AdminController_getAllAdmins",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List all administrators",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/delete-admin/{id}": {
      "delete": {
        "description": "Blocks the account from signing in and ends every session it has. The row and its role are kept — `orders.user_id` is ON DELETE RESTRICT, so hard-deleting anyone who has ordered would fail and would destroy order history. The id goes in the path; there is no request body. Refuses to target yourself, a non-administrator, or the last active super admin.",
        "operationId": "AdminController_deleteAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of the administrator to deactivate.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeactivatedDTO"
                }
              }
            }
          },
          "400": {
            "description": "That user is not an administrator."
          },
          "403": {
            "description": "Caller is not a super admin, or is targeting themselves."
          },
          "404": {
            "description": "No such user."
          },
          "409": {
            "description": "Already deactivated, or this is the last active super admin."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Deactivate an administrator",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/create-admin": {
      "post": {
        "description": "Creates a fresh administrator account. Buyers are never promoted into administrators — an account is created as one here, or by the `admin:create` CLI script that bootstraps the first super admin. `role` defaults to admin and may also be super_admin; buyer is rejected. The new account signs in at POST /auth/admin/login and has no Google or Facebook sign-in.",
        "operationId": "AdminController_createAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCreatedDTO"
                }
              }
            }
          },
          "400": {
            "description": "Missing field, password too short, or a non-administrator role."
          },
          "403": {
            "description": "Caller is not a super admin."
          },
          "409": {
            "description": "Username or email already taken."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new administrator",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "Clothing Store API",
    "description": "Authentication, catalogue and ordering endpoints. Generated from the controllers, so it cannot drift from the code.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "httpOnly refresh cookie, scoped to /auth. Set at sign-in and rotated by POST /auth/refresh."
      }
    },
    "schemas": {
...
      "AdminSummaryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "userName": {
            "type": "object",
            "example": "boss",
            "nullable": true
          },
          "email": {
            "type": "string",
            "example": "boss@example.com"
          },
          "firstName": {
            "type": "object",
            "example": "Site",
            "nullable": true
          },
          "lastName": {
            "type": "object",
            "example": "Owner",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "super_admin"
            ],
            "example": "super_admin"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "False once deactivated — the account cannot sign in."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T09:12:44.000Z"
          }
        },
        "required": [
          "id",
          "userName",
          "email",
          "firstName",
          "lastName",
          "role",
          "isActive",
          "createdAt"
        ]
      },
      "AdminListDTO": {
        "type": "object",
        "properties": {
          "admins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSummaryDTO"
            }
          },
          "total": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "admins",
          "total"
        ]
      },
      "AdminDeactivatedDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42
          },
          "userName": {
            "type": "object",
            "example": "olddesk",
            "nullable": true
          },
          "email": {
            "type": "string",
            "example": "old@example.com"
          },
          "role": {
            "type": "string",
            "example": "admin",
            "description": "Unchanged — deactivating does not alter the role."
          },
          "isActive": {
            "type": "boolean",
            "example": false
          },
          "sessionsRevoked": {
            "type": "number",
            "example": 2,
            "description": "Sessions ended by the deactivation, across all devices."
          }
        },
        "required": [
          "id",
          "userName",
          "email",
          "role",
          "isActive",
          "sessionsRevoked"
        ]
      },
      "CreateAdminDTO": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "example": "newadmin",
            "description": "Sign-in name. Must be unique."
          },
          "email": {
            "type": "string",
            "example": "newadmin@example.com",
            "description": "Must be unique."
          },
          "password": {
            "type": "string",
            "example": "pass12",
            "description": "At least 6 characters. Hashed with bcrypt before storage."
          },
          "firstName": {
            "type": "string",
            "example": "New"
          },
          "lastName": {
            "type": "string",
            "example": "Admin"
          },
          "phoneNumber": {
            "type": "string",
            "example": "admin"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "super_admin"
            ],
            "default": "admin",
            "description": "Defaults to admin. A super admin may also create another super_admin; buyer is rejected — use registration for those."
          }
        },
        "required": [
          "userName",
          "email",
          "password"
        ]
      },
      "AdminCreatedDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42
          },
          "userName": {
            "type": "object",
            "example": "newadmin"
          },
          "email": {
            "type": "string",
            "example": "newadmin@example.com"
          },
          "firstName": {
            "type": "object",
            "example": "New",
            "nullable": true
          },
          "lastName": {
            "type": "object",
            "example": "Admin",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "super_admin"
            ],
            "example": "admin"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-06T18:17:56.193Z"
          }
        },
        "required": [
          "id",
          "userName",
          "email",
          "firstName",
          "lastName",
          "role",
          "isActive",
          "createdAt"
        ]
      }
    }
  }
}